Move trimDecimals function into utils/wallet
